requirements.txt
Django[argon2]>=1.11,<1.12
django-users2>=0.2.1
django-environ>=0.4.0
djangorestframework
PyMySQL>=0.7.10
Pillow
django-material>=1.1.0<1.2
#django-reversion>=2.0.9
django-enumfields>=0.9.0


#django-imagekit
#drf-enum-field
#django-geoposition>=0.3.0
#captain-pycard>=0.9.9
#django-unixdatetimefield
#PyJWT>=1.4.2
#django-cors-headers
#pycrypto>=2.6.1
#gevent>=1.1.2
#gunicorn>=19.6.0
#psycopg2>=2.6.2
#honcho>=0.7.1
#whitenoise>=3.2.1
#raven>=5.26.0
#requests>=1.11.1
#psycogreen>=1.0
#Django[bcrypt]>=3.1.3
#Django>=1.11,<1.12
#django-admin-interface>=0.1.9
#celery
#django-channels
#django-import-export
#django-controlcenter
#djangorestframework
#django-cors-headers
#djangorestframework-jwt
#django-oauth-toolkit
#django-rest-swagger
#django-filter
#django-adminplus
#django-admin-tools
#django-mptt
#django-mptt-admin
#django-rest-auth
#django-admin-report
#django-daterange-filter
#django-allauth
#django-autocomplete-light
#django-select2
#django-smart-select
#django-imagekit
#django-grappelli
#djangocms-admin-style
#django-reversion
#django-debug-toolbar
#django-crispy-forms
#django-redis
#django-hystack
#django-taggit
#django-bootstrap3
#django-guardian
#django-push-notifications
#django-anymail
#sparkpost
#onesignal
#django-jenkins
#django-fsm
#django-constance[database]
#django-scheduler
#django-cacheops
#django-tables2
#django-rosetta
#django-ajax-selects
#django-storages
#django-minio-storage
#django-jet
